backend/requirements.txt
google
google.genai
pydantic
fastapi
uvicorn[standard]
python-multipart
aiofiles
python-dotenv
requests
pillowpyt